backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
PyMuPDF==1.19.6
werkzeug==2.0.1
Pillow==8.3.2 